:recycle: recruit requirement move to event
