+++ frontend/src/views/components/event/RecruitRequirements.tsx
